# src/main/java/com/dhr/chapatumoney/dto/response/FollowingResponse.java
package com.dhr.chapatumoney.dto.response;

import lombok.Builder;
import lombok.Data;

import java.time.OffsetDateTime;
import java.util.UUID;

@Data
@Builder
public class FollowingResponse {
    private UUID userId;
    private UUID artistId;
    private OffsetDateTime followedAt;
}